backend/agents/parser/requirements.txt
confluent-kafka
minio
pytesseract
pdf2image
PyPDF2
Pillow
redis
python-dotenv
fastapi
uvicorn
